+++ pom.xml
@@ -8,11 +8,11 @@
     <artifactId>LearningRestAssured</artifactId>
     <version>1.0-SNAPSHOT</version>
 
-    <!--    <properties>-->
-    <!--        <maven.compiler.source>11</maven.compiler.source>-->
-    <!--        <maven.compiler.target>11</maven.compiler.target>-->
-    <!--        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>-->
-    <!--    </properties>-->
+    <properties>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
 
     <dependencies>
         <!-- https://mvnrepository.com/artifact/io.rest-assured/rest-assured -->
@@ -101,7 +101,7 @@
         <dependency>
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-surefire-plugin</artifactId>
-            <version>3.0.0-M7</version>
+            <version>3.1.2</version>
         </dependency>
         <dependency>
             <groupId>io.rest-assured</groupId>
@@ -110,16 +110,4 @@
         </dependency>
 
     </dependencies>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>7</source>
-                    <target>7</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
 </project>
